Address review comments + reorganize packages a bit
