+++ compatibility-tests/multiclient/src/compatibility-test/java/tech/pegasys/artemis/compatibility/multiclient/clients/Prysm.java
@@ -30,7 +30,7 @@ import org.apache.tuweni.bytes.Bytes;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.utility.MountableFile;
-import tech.pegasys.artemis.networking.p2p.jvmlibp2p.LibP2PNodeId;
+import tech.pegasys.artemis.networking.p2p.libp2p.LibP2PNodeId;
 import tech.pegasys.artemis.networking.p2p.peer.NodeId;
 
 public class Prysm extends GenericContainer<Prysm> implements BeaconChainNode {
