+++ networking/eth2/src/main/java/tech/pegasys/artemis/networking/eth2/peers/Eth2PeerManager.java
@@ -24,8 +24,8 @@ import org.jetbrains.annotations.NotNull;
 import tech.pegasys.artemis.networking.eth2.rpc.beaconchain.BeaconChainMethods;
 import tech.pegasys.artemis.networking.eth2.rpc.beaconchain.methods.StatusMessageFactory;
 import tech.pegasys.artemis.networking.eth2.rpc.core.RpcMethods;
-import tech.pegasys.artemis.networking.p2p.jvmlibp2p.LibP2PNodeId;
-import tech.pegasys.artemis.networking.p2p.jvmlibp2p.PeerHandler;
+import tech.pegasys.artemis.networking.p2p.libp2p.LibP2PNodeId;
+import tech.pegasys.artemis.networking.p2p.network.PeerHandler;
 import tech.pegasys.artemis.networking.p2p.peer.NodeId;
 import tech.pegasys.artemis.networking.p2p.peer.Peer;
 import tech.pegasys.artemis.storage.ChainStorageClient;
