+++ networking/p2p/src/main/java/tech/pegasys/artemis/networking/p2p/libp2p/PeerManager.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.networking.p2p.jvmlibp2p;
+package tech.pegasys.artemis.networking.p2p.libp2p;
 
 import static tech.pegasys.artemis.util.alogger.ALogger.STDOUT;
 
@@ -31,6 +31,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
 import org.jetbrains.annotations.NotNull;
+import tech.pegasys.artemis.networking.p2p.network.PeerHandler;
 import tech.pegasys.artemis.networking.p2p.peer.NodeId;
 import tech.pegasys.artemis.networking.p2p.peer.Peer;
 import tech.pegasys.artemis.storage.ChainStorageClient;
@@ -112,5 +113,9 @@ public class PeerManager implements ConnectionHandler {
 
   public Stream<Peer> streamPeers() {
     return connectedPeerMap.values().stream();
+  }
+
+  public int getPeerCount() {
+    return connectedPeerMap.size();
   }
 }
