+++ networking/p2p/src/main/java/tech/pegasys/artemis/networking/p2p/libp2p/gossip/AggregateTopicHandler.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.networking.p2p.jvmlibp2p.gossip;
+package tech.pegasys.artemis.networking.p2p.libp2p.gossip;
 
 import static tech.pegasys.artemis.datastructures.util.AttestationUtil.get_indexed_attestation;
 import static tech.pegasys.artemis.datastructures.util.AttestationUtil.is_valid_indexed_attestation;
