+++ networking/p2p/src/main/java/tech/pegasys/artemis/networking/p2p/libp2p/gossip/GossipMessageHandler.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.networking.p2p.jvmlibp2p.gossip;
+package tech.pegasys.artemis.networking.p2p.libp2p.gossip;
 
 import com.google.common.eventbus.EventBus;
 import com.google.common.eventbus.Subscribe;
