+++ networking/p2p/src/main/java/tech/pegasys/artemis/networking/p2p/network/NetworkConfig.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.networking.p2p;
+package tech.pegasys.artemis.networking.p2p.network;
 
 import io.libp2p.core.crypto.PrivKey;
 import java.util.List;
