+++ networking/p2p/src/main/java/tech/pegasys/artemis/networking/p2p/network/PeerHandler.java
@@ -11,8 +11,12 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.networking.p2p.jvmlibp2p;
+package tech.pegasys.artemis.networking.p2p.network;
 
-import io.libp2p.core.multistream.ProtocolBinding;
+import tech.pegasys.artemis.networking.p2p.peer.Peer;
 
-public interface Protocol<T> extends ProtocolBinding<T> {}
+public interface PeerHandler {
+  void onConnect(final Peer peer);
+
+  void onDisconnect(final Peer peer);
+}
