+++ networking/p2p/src/main/java/tech/pegasys/artemis/networking/p2p/network/Protocol.java
@@ -11,12 +11,9 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.networking.p2p.jvmlibp2p;
+package tech.pegasys.artemis.networking.p2p.network;
 
-import tech.pegasys.artemis.networking.p2p.peer.Peer;
+import io.libp2p.core.multistream.ProtocolBinding;
 
-public interface PeerHandler {
-  void onConnect(final Peer peer);
-
-  void onDisconnect(final Peer peer);
-}
+// TODO - flesh out this interface and remove dependency on libp2p
+public interface Protocol<T> extends ProtocolBinding<T> {}
