+++ networking/p2p/src/test/java/tech/pegasys/artemis/networking/p2p/libp2p/gossip/AttestationTopicHandlerTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.networking.p2p.jvmlibp2p.gossip;
+package tech.pegasys.artemis.networking.p2p.libp2p.gossip;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
